@@ -1,10 +1,6 @@
 {
-  "name": "ng-books",
+  "name": "matsearch",
   "version": "0.0.0",
-  "license": "MIT",
-  "engines": {
-    "node": ">=6.9.1"
-  },
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -15,56 +11,51 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.3.1",
-    "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "^4.2.4",
-    "@angular/compiler": "^4.2.4",
-    "@angular/core": "^4.2.4",
-    "@angular/forms": "^4.2.4",
-    "@angular/http": "^4.2.4",
-    "@angular/material": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^4.2.4",
-    "@angular/platform-browser-dynamic": "^4.2.4",
-    "@angular/platform-server": "^4.3.1",
-    "@angular/router": "^4.2.4",
+    "@angular/animations": "~7.2.0",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "~7.2.0",
+    "@angular/compiler": "~7.2.0",
+    "@angular/core": "~7.2.0",
+    "@angular/forms": "~7.2.0",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "~7.2.0",
+    "@angular/platform-browser-dynamic": "~7.2.0",
+    "@angular/router": "~7.2.0",
     "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^4.1.0",
-    "@ngrx/router-store": "^4.1.0",
-    "@ngrx/store": "^4.1.0",
-    "@ngrx/store-devtools": "^4.0.0",
-    "bootstrap": "^4.0.0-beta",
-    "core-js": "^2.4.1",
-    "enhanced-resolve": "3.3.0",
+    "@ngrx/effects": "^7.4.0",
+    "@ngrx/router-store": "^7.4.0",
+    "@ngrx/store": "^7.4.0",
+    "@ngrx/store-devtools": "^7.4.0",
+    "bootstrap": "^4.3.1",
+    "core-js": "^2.5.4",
+    "enhanced-resolve": "^4.1.0",
     "font-awesome": "^4.7.0",
-    "immutable": "^3.8.2",
-    "jquery": "^3.0.0",
-    "ng-pwa-tools": "^0.0.15",
-    "npm": "^5.7.1",
-    "popper.js": "^1.11.0",
-    "rxjs": "^5.5.7",
-    "typescript": "~2.3.3",
-    "zone.js": "^0.8.12"
+    "immutable": "^4.0.0-rc.12",
+    "jquery": "^3.4.1",
+    "popper.js": "^1.15.0",
+    "rxjs": "~6.3.3",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.3.0-beta.0",
-    "@angular/compiler-cli": "^4.2.4",
-    "@angular/language-service": "^4.2.4",
-    "@types/angular-material": "^1.1.58",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.1.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "~7.3.8",
+    "@angular/compiler-cli": "~7.2.0",
+    "@angular/language-service": "~7.2.0",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.0.4",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.2.2"
   }
 }
