@@ -1,8 +1,7 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { Book } from 'app/models/book.model';
+import { Book } from '../../models/book.model';
 
 @Component({
-  moduleId: module.id,
   selector: 'ngb-books-list',
   templateUrl: 'books-list.component.html',
   styleUrls: ['books-list.component.scss']
