@@ -1,14 +1,16 @@
-﻿import {  ActionReducerMap,MetaReducer } from '@ngrx/store';
+﻿import {  ActionReducerMap,MetaReducer,combineReducers} from '@ngrx/store';
 import { RouterReducerState, routerReducer } from '@ngrx/router-store';
+import {compose} from '@ngrx/core/compose';
 import { RouterStateUrl } from '../../shared/utils';
-import * as Auth from './auth';
-export interface State extends Auth.AuthState {
+import * as Auth  from './auth';
+import { Actions } from '../actions/auth/auth.action';
+export interface State extends Auth.State  {
     router: RouterReducerState<RouterStateUrl>
 }
-export const reducer: ActionReducerMap<State> = {
+
+ export const reducer: ActionReducerMap<State> = {
     router: routerReducer,
-    ...Auth.Authreducer
-    //fromAuth: Auth.Authreducer.fromAuth,
-    //login: Auth.Authreducer.login
+    fromAuth:Auth.authReducer,
+    login:Auth.loginReducer
 };
 export const metaReducers: MetaReducer<State>[] = [];
